materialize: log progress of requesting Open
Getting the Open response for a materialization can take a long time, especially
if it has to do a lot of work to setup the transactor or if querying a
checkpoint from the store is not fast. This adds some progress logging for that
case so it's more clear what the materialization is doing.

Also a final round of refactoring: Split out the stream/logging logic into
separate files, and added a general `repeatAsync` helper that several of these
new logs make use of.
